devide store.js into other files
