+++ src/App.js
@@ -3,7 +3,8 @@ import MyForm from './components/MyForm'
 import ToggleButton from './components/ToggleButton'
 import TodoList from './components/TodoList'
 import v4 from 'uuid/v4'
-import store, { addTodoAction, toggleTodoAction } from './store'
+import store from './store'
+import { addTodoAction, toggleTodoAction } from './actions'
 
 class App extends Component {
   constructor(props) {
