+++ src/actions/_actionTypes.js
@@ -1,0 +1,4 @@
+export default {
+  ADD_TODO: 'ADD_TODO',
+  TOGGLE_TODO: 'TOGGLE_TODO',
+}
