+++ src/reducers/todoReducer.js
@@ -1,0 +1,7 @@
+export default (state, action) => {
+  if (state.id !== action.payload.id) return state
+  return {
+    ...state,
+    completed: !state.completed
+  }
+}
